- Fix to top() to get it right more often
- Make sure to catch exceptions when trying
  to reference "global" which throws on some strict
  platforms
- Better default path using the ESM "globalThis" if
  everything else fails
- Reorganized the tests so that browser and node based tests are
  in different files. That way, the browser can avoid running the
  node tests and node can avoid running the browser tests.
